backend/requirements.txt
fastapi
uvicorn
sqlalchemy
oracledb
pydantic
python-dotenv